Patch 0.2 FIX

+++ API/data/ban_list.py
@@ -8,7 +8,7 @@ from .db_session import SqlAlchemyBase
 class Ban_list(SqlAlchemyBase, SerializerMixin):
     __tablename__ = "Ban_list"
     id = sqlalchemy.Column(sqlalchemy.Integer, primary_key=True, autoincrement=True)
-    tg_id = sqlalchemy.Column(sqlalchemy.Integer, nullable=False)
+    tg_id = sqlalchemy.Column(sqlalchemy.BigInteger, nullable=False)
     class_id = sqlalchemy.Column(
         sqlalchemy.Integer, sqlalchemy.ForeignKey("classes.id"), nullable=False
     )
